Don't perform lookup before insertion
Previously, chash_insert performed a lookup in the hash table of
the data passed to it, and returned immediately pending a positive
result. Since this is a chained hash, we don't really care if the
same datum appears in the hash table more than once. Remove this
check. Also make miscellaneous changes otherwise, such as remove
unnecessary cast and specify `none' in an empty doc comment field.